Product Admin - securing products and user_profile views

+++ products/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 
@@ -71,8 +72,13 @@ def get_product_detail(request, product_id):
     return render(request, 'products/product_detail.html', context)
 
 
+@login_required
 def add_product_admin(request):
     """ Add a product to the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners have the permission to add products.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = AddProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -92,8 +98,13 @@ def add_product_admin(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product_admin(request, product_id):
     """ Edit a product in the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners have the permission to edit products.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = AddProductForm(request.POST, request.FILES, instance=product)
@@ -116,8 +127,12 @@ def edit_product_admin(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product_admin(request, product_id):
     """ Delete a product from the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners have the permission to delete products.')
+        return redirect(reverse('home'))
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Product is deleted!')
